feat: add `transform` argument
Passing a function to the constructor allows you to avoid creating a function for every value passed to `push`, `concat`, or `map`. Instead, each value is passed into the given transform function, which can return any value (including a promise). Since the transform function is called within `_run`, it's only called when each value is dequeued. This makes it perfect for performing the same task for every value in the queue.